pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manastone"
version = "0.1.0"
description = "Manastone Autonomic Operations Layer for Unitree G1"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19",
    "websockets>=12.0",
    "numpy>=1.24",
    "xgboost>=2.0",
    "scikit-learn>=1.3",
    "anthropic>=0.30",
    "fastmcp>=0.4",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.10",
    "hypothesis>=6.100",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
select = ["E", "F", "UP", "B", "I"]
ignore = ["E501"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py310"]
